Don't throw IOException if the close(2) failure was due to ECONNRESET
This seems to be relatively unique to *BSD and causes problems with at
least Jetty 9.x.

Please see Issue #94 for details.

This doesn't need to be ported to other openjdk versions.  Newer versions
have changed the way socket channel closure works to ignore an IOException
on close.  This could potentially be removed if that is ever backported to
openjdk11

With this change in place no change in Tier 1 internal tests were observed.

Discussed with:	@bsdkurt